backend/requirements.txt
flask==3.1.1
flask-cors==6.0.1
python-dotenv==1.1.1
requests==2.32.4
duckduckgo-search==8.1.1
asgiref
werkzeug
beautifulsoup4==4.13.4
lxml==6.0.0
psutil==7.0.0
tavily-python==0.7.9
pymongo==4.13.2
bson==0.5.10
markupsafe==3.0.2
markdown==3.8.2
uvicorn==0.35.0
fastapi==0.116.0
blinker>=1.9.0
jinja2>=3.1.2
itsdangerous>=2.2.0
aiofiles>=23.1.0
httpx>=0.28.1
soupsieve>1.2
tiktoken<1,>=0.5.2
anyio<4.0.0,>=3.7.1
starlette<0.47.0,>=0.40.0
httpcore==1.*
regex>=2022.1.18
primp>=0.15.0